pyproject.toml
[project]
name = "internetshop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "alembic>=1.15.2",
    "apscheduler>=3.11.0",
    "asyncpg>=0.30.0",
    "bcrypt==4.0.1",
    "fastapi>=0.115.12",
    "greenlet>=3.1.1",
    "loguru>=0.7.3",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.8.1",
    "pydantic[email]>=2.11.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "python-jose>=3.4.0",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.0",
]
